dont fail when TEI root is missing (print warning) #598

+++ .github/actions/ParlaMintValidate/validate.sh
@@ -15,17 +15,21 @@ for parla in $(jq -r '.[]' <<< $1 ); do
   echo "Cleaning old sample files [$parla]"
   rm -f ${DATADIR}/ParlaMint-$parla/ParlaMint-*.{txt,tsv,conllu,vert}
 
-  ( Scripts/validate-parlamint.pl Schema ${DATADIR}/ParlaMint-$parla 2>&1 || echo "ERROR: validate-parlamint.pl exited with <> 0" ) \
-    | sed "s/^\(.*\)\(\berror\b\)/::error::\1\2/i" | tee $DIR/validate.log
+  if [ -f "${DATADIR}/ParlaMint-$parla/ParlaMint-$parla.xml" ] ; then
 
-  echo "Validating parla-CLARIN (TEI)"
-  java -jar /usr/share/java/saxon.jar -xi -xsl:Scripts/copy.xsl ${DATADIR}/ParlaMint-$parla/ParlaMint-$parla.xml > $TESTDIR/ParlaMint-$parla.xml
-  java -jar /usr/share/java/jing.jar Schema/parla-clarin.rng $TESTDIR/ParlaMint-$parla.xml| sed "s/^\(.*\)\(\berror\b\)/::error::\1\2/i" | tee $DIR/parla-clarin-validate-tei.log
+    ( Scripts/validate-parlamint.pl Schema ${DATADIR}/ParlaMint-$parla 2>&1 || echo "ERROR: validate-parlamint.pl exited with <> 0" ) \
+      | sed "s/^\(.*\)\(\berror\b\)/::error::\1\2/i" | tee $DIR/validate.log
 
-  echo "CONVERT to text and metadata"
-  ( Scripts/parlamintp-tei2text.pl ${DATADIR}/ParlaMint-$parla $DIR 2>&1 || echo "ERROR: parlamintp-tei2text.pl exited with <> 0" ) \
-    | sed "s/^\(.*\)\(\berror\b\)/::error::\1\2/i" | tee $DIR/text.log
+    echo "Validating parla-CLARIN (TEI)"
+    java -jar /usr/share/java/saxon.jar -xi -xsl:Scripts/copy.xsl ${DATADIR}/ParlaMint-$parla/ParlaMint-$parla.xml > $TESTDIR/ParlaMint-$parla.xml
+    java -jar /usr/share/java/jing.jar Schema/parla-clarin.rng $TESTDIR/ParlaMint-$parla.xml| sed "s/^\(.*\)\(\berror\b\)/::error::\1\2/i" | tee $DIR/parla-clarin-validate-tei.log
 
+    echo "CONVERT to text and metadata"
+    ( Scripts/parlamintp-tei2text.pl ${DATADIR}/ParlaMint-$parla $DIR 2>&1 || echo "ERROR: parlamintp-tei2text.pl exited with <> 0" ) \
+      | sed "s/^\(.*\)\(\berror\b\)/::error::\1\2/i" | tee $DIR/text.log
+  else
+    echo "::warning::skipping TEI version validation - missing corpus root file ParlaMint-$parla/ParlaMint-$parla.xml"
+  fi
 
   if [ -f "${DATADIR}/ParlaMint-$parla/ParlaMint-$parla.ana.xml" ] ; then
     echo "Validating parla-CLARIN (TEI.ana)"
@@ -41,7 +45,7 @@ for parla in $(jq -r '.[]' <<< $1 ); do
       | tee $DIR/conllu.log
 
   else
-    echo "::warning::skipping annotated version validation - missing corpus root file"
+    echo "::warning::skipping TEI.ana version validation - missing corpus root file ParlaMint-$parla/ParlaMint-$parla.ana.xml"
   fi
 
   echo "Move new files to ParlaMint-$parla"
